Bug#16587369: HANDLER FOR SQLEXCEPTION NOT CATCHING ERROR 1452 (FK CONSTRAINT) FOR UPDATE
Post push fix. 

In the first patch, restructuring a boolean expression in 
an if statement resulted in the function "end_bulk_insert()"
not being called in situations where it should. 

The fix is to split the if statement into two separate 
statements, making sure evaluation is done in the correct 
order. This also improves code readability.