WL#6441 Convert mysql_secure_installation         script to C
mysql_secure_installation script currently
stores the user supplied password in a temporary
option file. This option file is later passed to
mysql client, which executes the required SQL
commands.

Since, storing password in an option file is
considered insecure, the script can be converted
to a C++ program so that the program can connect
to the server directly and execute the specified
commands using C API (libmysql).