Fix bug in clusterj range query on hash-only primary key
This bug was introduced with revno 4551 which implements query ordering
Primary key was incorrectly allowed to be used as an index even though
there were no query terms that could use the index.

A new test case is added to test that range queries on hash-only primary keys
will not use the non-existent ordered index.