Follow up patch of Bug#17214191 MEMCACHED ADD/SET ARE SLOWER THAN INSERTS
Approved by Jimmy on IM.

This follow up patch fixed the trunk weekly test failure.
The previous patch triggered a legacy bug.
On Solaris, strtok_r has different behavior. That is, if the input
string is a empty string, the third in/out parameter will not change
after this function was called.
So, on Solaris, we can't use this out parameter to check the end
condition. We should use the return value to check it.