Bug #17001980 RQG_PERFORMANCE_SCHEMA STOPPED WITH A CRASH IN HA_INNOBASE::CLONE.
Problem:

In the function ha_innobase::clone(), it was asserted that a thread
cannot clone a table handler that is used by another thread. It was
also asserted that the original table handler and the cloned table
handler must belong to the same transaction. Both are wrong.

Solution:

It is OK for a thread to clone a table handler that is used by another
thread. The original table handler and the cloned table handler can
belong to different transactions.  So both asserts are removed.

rb#3066 approved by Marko