BUG#16498740 - CREATE/ALTER: INCORRECT ERROR MSG. SEPARATE 	       UNSUPPORTED OPS FROM OTHER FAILURES
BACKGROUND:
CREATE or ALTER table can fail due to unsupported options
OR if the user forgot to specify something in the command.
In this particular case in the bug report, ALTER table failed 
because the user forgot to specify tablespace for ndb during 
create table time.
While performing alter table, the current error message said 
"doesn't have this option" which can be misleading to the user.
The given option is supported but the user forgot to specify
tablespace during create time.
Instead of misleading error message we can print a useful error
message to the user.

ANALYSIS:
On failure of Alter table, HA_WRONG_CREATE_OPTION was invoked
which mapped to ER_ILLEGAL_HA in handler::print_error.
This change was done in Bug#13840553.
ER_ILLEGAL_HA can be misunderstood by user that storage engine
(in this case ndbcluster) does not support altering column 
storage from main memory to disk whereas the problem is 
user forgot to specify tablespace for ndb during create table.

FIX:
To fix this bug, a new handler error HA_MISSING_CREATE_OPTION
is added and it is returned when ALTER fails which is then 
mapped to newly introduced error code ER_MISSING_HA_CREATE_OPTION 
in handler::print_error() function.