Bug#17316731 ONLINE ALTER TABLE...ADD PRIMARY KEY LOGGING MODIFIES THE SOURCE TABLE
When logging the delete-marking of a record during
online ALTER TABLE...ADD PRIMARY KEY
InnoDB must write the transaction ID to the log as it was before the
deletion or delete-marking of the record. When doing this, it
accidentally also overwrites the DB_TRX_ID field in the original
table. This could lead to locking bugs.

The buggy function row_log_table_delete() is being called by DELETE,
UPDATE (of primary key), and ROLLBACK. The bug only affects
ADD PRIMARY KEY, not other forms of online ALTER TABLE.

This bug has been there from the start of WL#6255 (the
table-rebuilding variant of online ALTER TABLE).

Approved on IM by Kevin Lewis