Bug#17279580:RPL.RPL_TYPECONV FAILED ON PB2 WITH SQL SYNTAX ERROR
Issue: check_type.inc encloses the error in "" which conflicts with the            
error text after the change in the error message due to WL#3656.                   
                                                                                   
Bug#17286464    I_RPL.RPL_OLD_DECIMAL_TO_NEW_DECIMAL FAILING ON                   
PB2 WITH ERROR IN SQL SYNTAX                                                       
                                                                                   
Issue: wait_for_slave_sql_error.inc encloses the error in "" which                 
conflicts with the  error text after the change in the error message               
due to WL#3656.                                                                    
                                                                                   
Fix: Removed the double quote in the error message since it will                
break existing applications trying to work on the error message.