WL#6982 Make plugin column in mysql.user non-null
Since the introduction of the post-4.1 authentication method,
the choice of authentication method was performed on the length
of the password hash alone, thus enabling easy change of the
authentication methods through a password change.

This has become a problem with the introduction of other
authentication methods in 5.5. We needed a special value (empty)
of the authentication plugin column and a set of special cases
throughout the code to support the backward compatible behavior.

Since in 5.6 the pre-4.1 password hash is to be deprecated this
worklog aims to remove this backward compatible layer and simplify
the authentication code by removing the special cases.

Now the native authentication methods will be subject to the same
rules as the other authentication methods.