Support clusterjpa use of cached JPQL queries
This fixes a bug where the second identical query uses a cached query.
The symptom is the use of em.createQuery with the same query string,
which results in OpenJPA passing openjpa.prepared.SQL as the query language
to NdbOpenJPAStoreManager.newQuery.

Thanks to Serdyn du Toit for the report and suggested fix.

The implementation of the fix is adapted from OpenJPA JDBCStoreManager.