Bug#16775543:MASTER SHUTDOWN NEEDS TO WAIT FOR SEMI-SYNC ACK OR TIMEOUT BEFORE COMPLETING.
Problem:
=======
When MySQl is shutdown and semi sync is enabled, the master
doesn't wait for semi sync ack or timeout before shutting
down. When the master shutsdown, it stops semi sync
reverting to async and whatever connections existed, and
whatever open transactions exist are allowed to finish.


Analysis:
========
During shutdown process all the active threads will be
informed to go down. In the existing code first client and
dump threads are informed to go down.  Because of this if
the dump thread has some pending events from active clients
which needs to be sent to slave they will not be sent.
Hence at the time of shutdown master has more events than
slave.

Fix:
===
Dump thread will be going down at the end.  First clients
will be informed to go down and then dump threads will be
stopped.