BUG 17305626 - MISSING TABLES IN INNODB DICTIONARY CAUSE ASSERTION AND RESTART OF MYSQL
The root cause of the problem is we will assert false in fil_node_open_file
when the *.ibd data files are missing.

The solution for the bug would be that when fil_node_open_file find *.ibd
data files are missing, do not assert false but return false, the caller
should deal with the return message. The caller of fil_node_open_file should
handle the new return message accordingly, unless it's too complicated or
no need to handle them. In the later case, assertion would be used.

fil_node_open_file also has several other assertions for the case that .ibd
data exists but has error data in itself, which should be distinguished from
.ibd data missing. So they're not handled in this patch.

Approved by Jimmy, rb#3129