Bug#17332162 UNNECESSARY DATA TYPE ENUM TRX_RB_CTX
In InnoDB transaction rollback, we pass a flag to the function that deletes a
record, to determine if the off-page columns should be freed.
It would suffice to pass a Boolean flag instead of the enum trx_rb_ctx that
we are currently passing.

See Bug#16971045 ASSERTION FAILURES ON ROLLBACK OF AN INSERT AFTER A FAILED
BLOB WRITE why it is unnecessary to make a distinction between the rollback of a
crash-recovered transaction and the rollback of an active transaction.

rb#3131 approved by Jimmy Yang