Bug#17049537: ASSERT 'M_STATUS == DA_ERROR' FAILED IN               DIAGNOSTICS_AREA ON TRIGGER + KILL QUERY
The problem was that if a statement was killed while executing
a trigger, an assert could be triggered. The bug had no effect
on release builds.

Triggers are implemented as stored routines. When a stored
routine ends with a failure, we copy error condition
information from the diagnostics area used during the stored
routine to the diagnostics area used by the caller. This assert
checks that the source diagnostics area in fact contains an
error.

The problem was that in the case of kill, an error might
not have been reported yet at this point. This patch fixes 
the bug by checking if an error condition is present before
trying to copy it.

This bug was a regression introduced by WL#6406.