Bug#17234723: Shortcut execution path in JOIN::optimize()
This is a partial fix for bug no. 68825/16598320:
Performance regressions for single-threaded workloads.

This patch shortcuts processing for some special cases, by skipping
some functions calls. Shortcuts are made based on the following conditions:

 - plan_is_const()
 - materialized_table_count == 0
 - partitioned_table_count == 0
 - outer_join == 0
 - select_lex->sj_nests.elements == 0
 - unit->item == 0 (query block is not a subquery)

Some loops are also made into functions, so that branching in case the
function is skipped is minimized, thus avoiding cache misses.

The gain is up to 5% on a simple synthetic benchmark (see benchmark
BM-1 in bug 68825 for a description), but notice that the expected gain
is platform-dependent.