BUG#14767986: START SLAVE UNTIL SQL_AFTER_GTIDS NEEDS UNTIL_CONDITION_SET+1 GTID TO STOP
Problem:
=======
START SLAVE UNTIL SQL_AFTER_GTIDS needs
until_condition_set+1 GTID to stop. This is a problem since
we may not have the next GTID event, or the next GTID event
is delayed.

SQL_BEFORE_GTID case is also flawed since it does not stop
the slave thread if the GTID being waited for is already
applied

Fix:
===
We check if the until condition is satisfied before the sql
thread goes on a wait for the relay log to grow. We also
make the slave thread to stop in case the GTID being waited
for in SQL_BEFORE_GTID has already been applied.

The above fix will still fail for SQL_AFTER_GTIDS when MTS
is enabled. This will be fixed in future. Hence when MTS is
enabled along with SQL_AFTER_GTIDS following warning will
ER_MTS_FEATURE_IS_NOT_SUPPORTED be thrown and slave will be
switched to single threaded mode.