Bug#17315083 - SUPPORT INTEGER KEY/VALUE FOR MEMCACHED TABLE Bug#17203937 THE MEMCACHED PLUGIN CAN BE INITIALIZED WITH AN INTEGER PK.
Approved by Jimmy on rb://3082.

This patch is for fixing 2 bugs, bug#17203937 and bug#17315083.

In the bug17203937, the reason that the memcached plugin still can be
initialized is, there's a correct config row in containers. This config
row is inserted by the innodb_memcache_config.sql script.
We should allow the plugin can be installed, even if there're some of
rows in containers are not correct.
For the not correct rows in containers, we need print some information
about what's wrong with it.

For bug17315083, we add support for integer columns. Not only the key
column can be int, the value column can be int also. The code of this
part is based on Jimmy's patch on rb://3083.