Fix Bug#16511145 INNODB TOO SLOW AT CREATING OR ALTER A TABLE TO ADD LARGE NO. OF INDEXES
During ALTER TABLE ADD INDEX save only the stats for the new index into
mysql.innodb_index_stats. Saving the stats for the rest of the indexes
is not necessary because they have not been changed.

Another change in this patch is to reduce the number of calls to
dict_stats_exec_sql() during stats save. The logic of the code is:
if the data is in the tables, then UPDATE, otherwise INSERT. Previously
we did this by first attempting an INSERT and if duplicate key error is
returned, then do the UPDATE. The thing is that for this we need two
calls to the internal SQL parser because we cannot check if INSERT
failed with duplicate key error from within one SQL procedure. Instead,
in this patch, we do "DELETE, INSERT" in one SQL procedure with one call
to dict_stats_exec_sql() instead of two.

Reviewed by:	Sunny (rb:3107)