Bug#17309863 AUTO RECONNECT DOES NOT WORK WITH 5.6 LIBMYSQLCLIENT Problem: A client program fails to reconnect to a mysql-5.6 server after a connection is killed by a KILL CONNECTION_ID() command. Any query executed after that returns "query failed, Lost connection to MySQL server during query". This is a regression introduced by a patch for Bug#11762221 (54790: USE OF NON-BLOCKING MODE FOR SOCKETS LIMITS PERFORMANCE)
Solution:
The patch is restoring the behavior of the previous versions.
In net_serv.cc::net_clear(), setting net->error=2 flag if 
connection was terminated, allows the client to reconnect.
We check if the peer is still connected (EOF not reached). 
Otherwise we set the error flag so that the connection might be
re-established before we try to send a query to the server.