BUG#17036976 - INCORECT ERROR FOR GRANT PROCEDURE
BACKGROUND:
GRANT for non-existing PROCEDURE returns error 
"FUNCTION or PROCEDURE ..."
even if FUNCTION with given name really exists, 
which is little bit misleading.
The same is true for non-existing FUNCTION which 
returns the same error even if PROCEDURE really
exists. This error can be confusing to the user.

ANALYSIS:
In sp_exist_routines(), if the routine is not 
found it returns this error.
Based on the value of is_procs flag (true for PROCEDURES
false for FUNCTIONS) we can distinguish
whether it is a PROCEDURE or a FUNCTION and we can 
return the appropriate error to the user.

FIX:
A small check has been added to properly distinguish 
between PROCEDURE or FUNCTION based on is_procs flag
to correctly pass the error message parameter in 
sp_exist_routines() from where the error is thrown.