This is patch for bug#17055378 - MULTI UPDATE DISPLAYS NO ERROR/WARNING FOR NOT NULL COLUMNS.
The reason for the bug was that temporary table's fields were marked as
temporary nullable without taking into account the presence of triggers for
a table. It led to erroneous possibility to set a NOT-NULL column of the table
being updated by multi-table UPDATE statement to NULL. This bug appeared
only when a table being updated was not the first in-place updatable table
in a list of tables of a statement UPDATE. For such tables we have to create
corresponding  temporary table where all changes will be made.