Bug#17399658 MISSING PSI_MEMORY_KEY WHEN BUILDING PLUGINS WITHOUT THE PERFORMANCE SCHEMA
This fix is a build issue.

When compiling without the performance schema,
the declaration and definition of some PSI_memory_key for the memory 
instrumentation did not match.

The declaration was always present (both with and without the performance 
schema), as intended.

The definition was only present when compiling with the performance schema,
leading to unresolved symbols when compiling without.

The fix is to always declare and define every keys.

Also, found and fixed a my_alloc() / free() mismatch.