Bug#16602002 PROVIDE OPTION TO SKIP NODEGROUP=65536 NODES FROM NDB_ERROR_REPORTER
Modified ndb_error_reporter to add a --connection-timeout parameter for 
a timeout while connecting to nodes, a --skip-nodegroup parameter to 
skip all data nodes belonging to a specific nodegroup, and a --dry-scp 
option to print the scp command instead of executing it.

Added a unit test for ndb_error_reporter in mtr.