Bug#17400029 - UNDERCOUNTED FREE OPERATIONS IN PERFORMANCE SCHEMA MEMORY INSTRUMENTATION
In the performance schema memory instrumentation,
statistics collected about FREE operations could be under evaluated,
leading to a perceived memory leak.

The problem was that when a memory block was counted as allocated,
it was not always counted as freed, because the logic in
memory_free_v1() and memory_realloc_v1() tested too many flags.

During a call to memory_free_v1(),
the value of:
- the global_instrumentation consumer
- the thread_instrumentation consumer
- the thread.INSTRUMENTED column
- the presence of thread instrumentation

may affect *how* free statistics are collected,
for example accounted per thread or globally,
but it should not affect *if* free statistics are collected.

With this fix, there is no logic affecting *if* free statistics
are collected.
Every memory block counted as allocated is then counted as freed,
when invoking the instrumentation for the free (or realloc) operation.

This fix also renamed the memory instrumentation apis.