Bug#17406891 - CRASH IN MDL INSTRUMENTATION, WHEN THREAD IS NOT INSTRUMENTED
In pfs_rebind_table_v1(),
the performance schema instrumentation for table handles
could crash, when assigning a table handle to a non instrumented thread.

With this fix,
for table performance_schema.table_handles,
column OWNER_EVENT_ID is set to 0
when the thread that owns a table handle is not instrumented.