BUG#16982071: ASSERT `TAB->KEYS.IS_SET(KEYNO)' FAILED IN               SETUP_SEMIJOIN_DUPS_ELIMINATION
Consider a query where SemiJoin LooseScan is used on a 
subquery. After deciding on join order, the range optimizer
is rerun in the "rechecking index usage" code block of 
make_join_select(). After that, tab->keys (which is the bitmap
of keys that are usable by the query, and is also used to print
"possible_keys" in EXPLAIN) is overwritten by 
SQL_SELECT::quick_keys under the assumption that the range
optimizer sets bits for all usable indexes in 
SQL_SELECT::quick_keys:

  if (!sel->quick_keys.is_subset(tab->checked_keys) ||
      !sel->needed_reg.is_subset(tab->checked_keys))
  {
    tab->keys=sel->quick_keys;  // HERE
    ...
  }

Unfortunately, the range optimizer never sets 
these bits, it sets the bits in the variable with the same 
name in TABLE::quick_keys instead. Thus, tab->keys is 
incorrectly cleared. When setup_semijoin_dups_elimination()
is called later, it fails on a DBUG_ASSERT() that is meant to 
verify that the index chosen for LooseScan is actually one 
that is applicable (i.e., asserts that the bit is set in 
tab->keys).

Since the SQL_SELECT variable never gets any bits set, it is 
removed by this patch to avoid code duplication and improve 
readability. The few locations that used to check for set 
bits (always 0) in this variable now check TABLE::quick_keys 
instead.

This commit also:
 - removes an unused enum value (SEL_ARGs 'MAYBE' type) from 
   opt_range.cc
 - adds info to optimizer trace: if the range optimizer cannot
   use an index because the range predicates depend on values
   from another table, the previously printed 
   "chosen: false, cause: unknown" is now printed as
   "chosen: false, cause: depends_on_unread_values".