Bug#17400687: DIAGNOSTICS_AREA::SET_ERROR_STATUS ASSERTION
This assert is triggered if a statement tries to set an exception
condition in a diagnostics area which already contains an
exception condition or a completion condition.

It was triggered in this case by a failed stored program trying
to transfer its exception condition to the diagnostics area
of the caller - which already contained an exception condition.
Normally this should not happen, but there is a special case
with multi update, non-transactional tables and triggers.

If a multi-update has already updated a non-transactional
table when it encounters an error updating another table,
it cannot rollback the changes. In this case, it tries to
proceed and may end up executing a trigger. If this trigger
also fails, the assert was triggered.

This patch fixes the problem by not copying exception condition
from the diagnostics area of the stored program to the diagnostics
area of the caller if the latter already contains an exception
condition.

This bug was (likely) a regression introduced by WL#6406.