Implement order, skip, and limit for queries
If used, the order parameter must be a string that is ignoreCase 'asc' or 
'desc'. Ordering is not explicit; rather, the filter must resolve to 
an index scan. The index is scanned either ascending or descending.

Use of skip or limit requires the order to be specified, either ascending 
or descending. Either or both of skip and limit may be specified. If only 
skip is specified, limit is MAX_INT (2^32-1). If only limit is specified, 
skip is 0. Skip is the number of rows to skip before returning limit rows.

MySQLConnection.js:
  Add a new ErrorOperation that does not execute any database operations.
  Instead, it is put into the execution framework. When it is executed,
  it calls back into UserContext to report the error, and then calls back
  into TransactionHandler.
  ErrorOperation is used to report errors during query execution but is
  generic so it can be used in other places to report certain errors.

QueryBtreeIndexScanLimitTest.js:
  new test for btree index scan order, skip, and limit

QueryPrimaryIndexScanLimitTest.js:
  new test for primary index scan order, skip, and limit

harness.js:
  add TestCase.errorIfNotError to report missing errors