Bug#16967281 ASSERTION FAILED: PARAM->FUNC_COUNT <= COPY_FUNC_COUNT
Problem: Queries with ROLLUP and an inner subquery with a reference to
an outer field that is replaced by JOIN::rollup_process_const_fields()
trigger an assertion in create_tmp_table().

After calling JOIN::rollup_process_const_fields(), some items in the
grouping list may be changed to Item_func_rollup_const. This may
affect the count of fields and functions in the temporary table, so
count_field_types() is called to do a recount.

However, a field that has changed may also be used in inner
subqueries, but these are not recounted. When execution gets to
create_tmp_table() for the inner subquery, the precalculated field and
function counts don't match what was seen by create_tmp_table(), so an
assertion is fired.

Fix: Add a new function JOIN::recount_field_types() that recursively
call count_field_types() on subqueries and call it after calling
JOIN::rollup_process_const_fields().