Bug#17420682: SQL_KILL: ASSERTION FAILED: ! IS_SET(),               FILE SQL_ERROR.CC, LINE 382
This assertion is triggered if a statement tries to send an OK
packet to the client after the statement has reported an error.

It was triggered in this case if KILL [CONNECTION | QUERY] was
used with a thread_id argument that somehow gave an error.
The implementation of KILL did not take this possibility into
account. This patch fixes the problem by checking if val_int()
has reported an error.

No regression test is included as the reported test first hits
the assertion bug reported as Bug 11764690: SOME FUNCTIONS CAUSE
ASSERTION WHEN GIVEN A ROW CONSTRUCTOR ARGUMENT. The patch was
verified by temporarily disabling the Bug 11764690 assert.