Bug#17420109: SETTING SLAVE_PARALLEL_TYPE=DEFAULT CRASH IN CHECK_UPDATE_MTS_TYPE
Problem: setting the value of slave-parallel-type variable on a
standalone server may cause the server to exit. On a server, not
configured as slave active_mi is null and accessing it may cause
failures.

Fix: fixed the issue by taking proper locks and checking if
active_mi is not null