Bug #17277744 ERROR 1777 IN CHANGE TO MASTER_AUTO_POSITION = 0 WHEN AUTO_POSITION =1
Problem: If Master-Slave are using GTID based
replication and the dump protocol was configured
to use GTIDs and if that slave is restarted with
gtid_mode=OFF, initialization on slave which is
required for replication is failing with error
ER_AUTO_POSITION_REQUIRES_GTID_MODE_ON(1777).
And any further commands like "start slave"
or "change master" are also throwing the same
ER_AUTO_POSITION_REQUIRES_GTID_MODE_ON(1777)
error.

Analysis: When the slave server is restarted,
it goes into initialization phase and reads
master info file. Slave server was throwing
error if it find Auto_position is 1 and
gtid_mode is off and stops rest of the
initialization required for replication.

Fix: Slave server should not throw error if
auto_position is 1 and gtid_mode is off.
Instead, Server forces(resets) the auto_position
value to 0.