Bug #16629195 	INVALID USE OF GRANT COMMAND BREAKS REPLICATION
To failed REVOKE/GRANT statement, we wrote an incident event
into binary log, which breaks replication on all slaves.

Before ACLs are changed to execute fully or none at all,
we go with the following fix, write an incident if one
or more users are granted/revoked successfully (it has
a partial execution), otherwise, do not write an incident,
just log a warning.