Bug #16764240 SELECT FROM INNODB TABLE HANGS BEING RUN IN PARALLEL WITH REPLICATING LOAD DATA
Problem:

The mtr_memo_contains() function determines if the given object is part
of the memo of the mini transaction.  In the test scenario, this search
is very slow (takes more than 1 hour).   This is because the search in
mtr_memo_contains() involves a nested loop.

Solution:

Re-write mtr_memo_contains() so that the search involves only one loop.
This is already done by Yasufumi as part of wl#6363.  Backporting the
functions mtr_memo_contains() and mtr_memo_note_modifications() from
mysql-5.7 to mysql-5.6 is sufficient. 

rb#3342 approved by Yasufumi