Bug #17258782 	MASTER DIDN'T WRITE ERROR MESSAGE IN THE LOG FILE FOR INCIDENT
Master is not generating error message to log file
when writing an incident event to binlog. So user
is not alerted that slave servers will be stopped
by the incident event later.

Master will write error messages in the log file
for all incidents. So that user might have a chance
to be alerted and explore incident details. Master
also make incident event carry the error message,
so that it can be displayed with the incident on
slave.