Bug#17478068 COMPILING THE PERFORMANCE SCHEMA WITH ONLY SOME INSTRUMENTATIONS
Before this fix, building the server could be done:
- with the performance schema, including ALL available instrumentation
- without the performance schema.

The problem is that given the growing list of available instrumentations,
this is less and less practical.

In particular, to isolate a functional or performance issue,
it can be desirable to build only with a particular instrumentation.

This fix introduce the following cmake options:
- DISABLE_PSI_MUTEX
- DISABLE_PSI_RWLOCK
- DISABLE_PSI_COND
- DISABLE_PSI_FILE
- DISABLE_PSI_TABLE
- DISABLE_PSI_SOCKET
- DISABLE_PSI_STAGE
- DISABLE_PSI_STATEMENT
- DISABLE_PSI_SP
- DISABLE_PSI_IDLE
- DISABLE_PSI_STATEMENT_DIGEST
- DISABLE_PSI_METADATA
- DISABLE_PSI_MEMORY
which can be used to not build with a specific instrumentation

When a specific instrumentation is not built:
- Instrumentation points added in the server code,
  which uses macros like PSI_MUTEX_CALL,
  do not make calls to the performance schema, saving CPU.
- The performance schema does not allocate memory for the instrumentation,
- The thread aggregation code does not maintain statistics for the
  instrumentation, as all statistics are 0, to save CPU.

Note that code like PFS_mutex_class is still compiled,
but effectively dead.
The goal is to reduce the memory and CPU footprint,
not to eliminate completely symbols from the binary.

Note that the table schema is unchanged.
Tables that depend on a specific instrumentation,
such as performance_schema.mutex_instances,
will be empty when building with DISABLE_PSI_MUTEX=ON,
but the table still exists.

These options are intended to be used primarily by maintainers,
and as such comes with the following limitation:

The test suite has _not_ been fixed to have each test check for additional
pre conditions, such as:
- instrumentation X must be available
- instrumentation Y must be available
to make the tests suite pass no matter what the combination of compiling option
is used.
Adding these tests would make the test suite hard to maintain, for no additional
value.

The only test expected to pass is perfschema.start_server_nothing,
which verifies that the server is still functional even with missing
instrumentation.